pyproject.toml
[project]
name = "super-biz-agent-py"
version = "1.2.1"
description = "基于 LangChain 的智能业务代理系统 - 支持 RAG 知识库和 AIOps 智能运维"
authors = [{name = "chief"}]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.1.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.0",
    "langchain-openai>=1.0.0",
    "langgraph>=0.0.40",
    "dashscope>=1.14.0",
    "openai>=1.10.0",
    "pymilvus>=2.3.5",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "langchain-milvus>=0.3.3",
    "langchain-text-splitters>=1.1.0",
    "langchain-mcp-adapters>=0.2.1",
    "fastmcp>=2.14.0",
    "langchain-qwq>=0.3.4",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "alibabacloud_ocr_api2021>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
    "ipdb>=0.13.13",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]



[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "logs",
    "*.pyc",
    "*.egg-info",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.isort]
known-first-party = ["app"]
combine-as-imports = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
split_on_trailing_comma = true
known_first_party = ["app"]
skip = [".venv", "venv", "__pycache__", "logs"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["dashscope.*", "langchain.*", "pymilvus.*"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101", "B601"]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*_test.py"]

[tool.pyright]
include = ["app"]
exclude = ["**/node_modules", "**/__pycache__", ".git", ".venv", "venv", "logs"]
pythonVersion = "3.11"
pythonPlatform = "Darwin"
typeCheckingMode = "basic"

reportMissingImports = "none"
reportMissingTypeStubs = "none"
reportUnknownParameterType = "none"
reportUnknownArgumentType = "none"
reportUnknownLambdaType = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportUntypedFunctionDecorator = "none"
reportUntypedClassDecorator = "none"
reportCallInDefaultInitializer = "none"
reportGeneralTypeIssues = "none"
reportOptionalMemberAccess = "none"

reportUnusedImport = "information"
reportUnusedVariable = "information"
reportDeprecated = "information"